# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Receipt Tracker API
  version: 0.1.0
  description: >
    API contract for a mobile-first receipt tracking app with OCR normalization,
    review, reporting, and outbound sync to QuickBooks Online and Microsoft Excel.
servers:
  - url: https://api.example.com/api/v1
    description: Production
  - url: http://localhost:8000/api/v1
    description: Local development
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    receiptId:
      name: receiptId
      in: path
      required: true
      schema:
        type: string
        format: uuid
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
              additionalProperties: true
    Pagination:
      type: object
      required: [limit, offset, total]
      properties:
        limit:
          type: integer
          minimum: 1
        offset:
          type: integer
          minimum: 0
        total:
          type: integer
          minimum: 0
...
    ReceiptLineItem:
      type: object
      required: [description]
      properties:
        description:
          type: string
        quantity:
          type: number
          nullable: true
        unitPrice:
          type: string
          nullable: true
        lineTotal:
          type: string
          nullable: true
    ReceiptFile:
      type: object
      required: [objectKey, mimeType]
      properties:
        objectKey:
          type: string
        mimeType:
          type: string
        originalFilename:
          type: string
          nullable: true
    ReceiptCreateRequest:
      type: object
      required: [file]
      properties:
        file:
          $ref: '#/components/schemas/ReceiptFile'
        source:
          type: string
          enum: [camera, upload, email, api]
          default: camera
        notes:
          type: string
          nullable: true
    FieldConfidence:
      type: object
      additionalProperties:
        type: number
        minimum: 0
        maximum: 1
    Receipt:
      type: object
      required:
        [id, status, source, currency, createdAt, updatedAt, syncTargets]
      properties:
        id:
          type: string
          format: uuid
        status:
          type: string
          enum:
            [uploaded, processing, review_required, approved, syncing, synced, failed, archived]
        source:
          type: string
          enum: [camera, upload, email, api]
        merchantName:
          type: string
          nullable: true
        transactionDate:
          type: string
          format: date
          nullable: true
        currency:
          type: string
          minLength: 3
          maxLength: 3
        subtotal:
          type: string
          nullable: true
        tax:
          type: string
          nullable: true
        tip:
          type: string
          nullable: true
        total:
          type: string
          nullable: true
        paymentMethod:
          type: string
          nullable: true
        categoryId:
          type: string
          format: uuid
          nullable: true
        categoryName:
          type: string
          nullable: true
        lineItems:
          type: array
          items:
            $ref: '#/components/schemas/ReceiptLineItem'
        notes:
          type: string
          nullable: true
        confidence:
          $ref: '#/components/schemas/FieldConfidence'
        syncTargets:
          type: object
          properties:
            quickbooks:
              type: string
              enum: [not_connected, not_requested, queued, syncing, synced, failed]
            excel:
              type: string
              enum: [not_connected, not_requested, queued, syncing, synced, failed]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ReceiptListResponse:
      type: object
      required: [data, pagination]
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Receipt'
        pagination:
          $ref: '#/components/schemas/Pagination'
    ReceiptUpdateRequest:
      type: object
      properties:
        merchantName:
          type: string
        transactionDate:
          type: string
          format: date
        subtotal:
          type: string
        tax:
          type: string
        tip:
          type: string
        total:
          type: string
        paymentMethod:
          type: string
        categoryId:
          type: string
          format: uuid
        lineItems:
          type: array
          items:
            $ref: '#/components/schemas/ReceiptLineItem'
        notes:
          type: string
    ReceiptApprovalRequest:
      type: object
      properties:
        approved:
          type: boolean
          default: true
        syncAfterApproval:
          type: array
          items:
            type: string
            enum: [quickbooks, excel]
    ReceiptSyncRequest:
      type: object
      required: [targets]
      properties:
        targets:
          type: array
          minItems: 1
          items:
            type: string
            enum: [quickbooks, excel]
        forceResync:
          type: boolean
          default: false
...
paths:
...
  /receipts:
    post:
      tags: [receipts]
      summary: Create a receipt record after upload
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReceiptCreateRequest'
      responses:
        '201':
          description: Receipt created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Receipt'
    get:
      tags: [receipts]
      summary: List receipts
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: q
          in: query
          schema:
            type: string
        - name: fromDate
          in: query
          schema:
            type: string
            format: date
        - name: toDate
          in: query
          schema:
            type: string
            format: date
        - name: categoryId
          in: query
          schema:
            type: string
            format: uuid
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: Receipt page
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReceiptListResponse'
  /receipts/{receiptId}:
    get:
      tags: [receipts]
      summary: Get a receipt
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/receiptId'
      responses:
        '200':
          description: Receipt details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Receipt'
        '404':
          description: Receipt not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    patch:
      tags: [receipts]
      summary: Update approved or suggested receipt fields
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/receiptId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReceiptUpdateRequest'
      responses:
        '200':
          description: Receipt updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Receipt'
  /receipts/{receiptId}/approve:
    post:
      tags: [receipts]
      summary: Mark a receipt as reviewed and ready for sync
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/receiptId'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReceiptApprovalRequest'
      responses:
        '200':
          description: Receipt approved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Receipt'
  /receipts/{receiptId}/retry-processing:
    post:
      tags: [receipts]
      summary: Retry OCR and normalization for a failed receipt
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/receiptId'
      responses:
        '202':
          description: Retry queued
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: queued
  /receipts/{receiptId}/sync:
    post:
      tags: [receipts]
      summary: Queue outbound sync jobs for a receipt
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/receiptId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReceiptSyncRequest'
      responses:
        '202':
          description: Sync queued
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: queued
...